nacl expects xlen to be a uint64, fixes #1

+++ dht.c
@@ -89,7 +89,7 @@ static int fd = -1;
 #pragma mark utils
 
 void
-randombytes(uint8_t *x, uint32_t xlen) {
+randombytes(uint8_t *x, uint64_t xlen) {
   int i;
   if(fd == -1) {
     for(;;) {
